[Correção de entidades.]
