+++ br.com.gps_2019_1/entidade/Pesquisa.java
@@ -1,5 +1,7 @@
 package entidade;
 
+import java.time.LocalDate;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
@@ -17,6 +19,10 @@ public class Pesquisa extends Entidade{
 	@ManyToOne
 	@JoinColumn(name = "usuario_id")
 	private Usuario usuario;
+	@Column(name = "data_inicio")
+	private LocalDate dataInicio;
+	@Column(name = "data_fim")
+	private LocalDate dataFim;
 	
 	public String getTitulo() {return titulo;}
 	public String getDescricao() {return descricao;}
@@ -34,5 +40,30 @@ public class Pesquisa extends Entidade{
 	public void setUsuario(Usuario usuario) {
 		this.usuario = usuario;
 	}	
+	
+	/**
+	 * @return the dataInicio
+	 */
+	public LocalDate getDataInicio() {
+		return dataInicio;
+	}
+	/**
+	 * @param dataInicio the dataInicio to set
+	 */
+	public void setDataInicio(LocalDate dataInicio) {
+		this.dataInicio = dataInicio;
+	}
+	/**
+	 * @return the dataFim
+	 */
+	public LocalDate getDataFim() {
+		return dataFim;
+	}
+	/**
+	 * @param dataFim the dataFim to set
+	 */
+	public void setDataFim(LocalDate dataFim) {
+		this.dataFim = dataFim;
+	}
 	
 }
